Clean up fixed-buffer sign comparisons
Cast fixed local buffer and table bounds to the existing signed loop
index type in a few parser and boot-loader loops. These bounds are known
small constants, so this clears -Wsign-compare without changing behavior.

+++ simulators/ALTAIR/altair_sys.c
@@ -252,7 +252,7 @@ if ((sw & SWMASK ('C')) || ((*cptr == '"') && cptr++)) { /* ASCII string? */
    or numeric (including spaces).
 */
 
-while (i < sizeof (gbuf) - 4) {
+while (i < (int32)(sizeof (gbuf) - 4)) {
     if (*cptr == ',' || *cptr == '\0' ||
          sim_isdigit(*cptr))
             break;
