+++ simulators/Intel-Systems/common/i8008.c
@@ -1152,7 +1152,7 @@ if ((sw & SWMASK ('C')) || ((*cptr == '"') && cptr++)) { /* ASCII string? */
 
 /* An instruction: get opcode (all characters until null, comma,
    or numeric (including spaces). */
-while (i < sizeof (gbuf) - 4) {
+while (i < (int32)(sizeof (gbuf) - 4)) {
     if (*cptr == ',' || *cptr == '\0' ||
          sim_isdigit(*cptr))
             break;
